requirements.txt
accelerate==0.33.0
datasets==2.21.0
decord==0.6.0
deepspeed==0.13.1
diffusers==0.30.0
einops==0.8.0
ftfy==6.2.3
gradio==3.44.0
imageio==2.35.1
iopath==0.1.9
joblib==1.3.1
matplotlib==3.9.2
numpy==1.26.3
pandas==2.2.2
peft==0.12.0
Pillow==10.2.0
pytorchvideo==0.1.5
PyYAML==6.0.2
regex==2024.7.24
scipy==1.14.2
timm==1.0.8
torch==2.1.2
torchaudio==2.1.2
torchvision==0.17.1+cu121
transformers==4.42.0
tensorboard==2.17.1
sacrebleu==1.5.0
sacremoses==0.0.46
webrtcvad==2.0.10
pydub==0.25.1
wandb==0.12.9
SoundFile==0.10.3.post1
PyYAML==6.0
scikit_learn==1.0.2
tweepy==4.5.0
sentencepiece==0.1.96